requirements.txt
requests==2.32.3
python-dotenv==1.0.1
shapely==2.0.6
geopandas==1.0.1
pandas==2.2.3
folium==0.19.4
branca==0.8.1
tqdm==4.67.1
